requirements.txt
fastapi[standard]
uvicorn
google-generativeai
python-dotenv
pydantic
